Adding environment configuration to TestFunctionHost
